+++ src/main/java/edu/ie3/datamodel/io/source/csv/CsvWeatherSource.java
@@ -41,6 +41,7 @@ import org.locationtech.jts.geom.Point;
 
 /** Implements a WeatherSource for CSV files by using the CsvTimeSeriesSource as a base */
 public class CsvWeatherSource extends WeatherSource {
+  private final Map<Point, IndividualTimeSeries<WeatherValue>> coordinateToTimeSeries;
 
   private final CsvDataSource dataSource;
 
@@ -64,7 +65,7 @@ public class CsvWeatherSource extends WeatherSource {
       throws SourceException {
     super(idCoordinateSource, weatherFactory);
     this.dataSource = new CsvDataSource(csvSep, folderPath, fileNamingStrategy);
-    coordinateToTimeSeries = getWeatherTimeSeries();
+    this.coordinateToTimeSeries = getWeatherTimeSeries();
   }
 
   // -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
@@ -96,6 +97,12 @@ public class CsvWeatherSource extends WeatherSource {
     IndividualTimeSeries<WeatherValue> timeSeries = coordinateToTimeSeries.get(coordinate);
     if (timeSeries == null) return Optional.empty();
     return timeSeries.getTimeBasedValue(date);
+  }
+
+  @Override
+  public Map<Point, List<ZonedDateTime>> getTimeKeysAfter(ZonedDateTime time) {
+    return coordinateToTimeSeries.entrySet().stream()
+        .collect(Collectors.toMap(Map.Entry::getKey, t -> t.getValue().getTimeKeysAfter(time)));
   }
 
   // -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-