+++ src/main/java/edu/ie3/datamodel/io/source/sql/SqlWeatherSource.java
@@ -41,6 +41,8 @@ public class SqlWeatherSource extends WeatherSource {
 
   private final String queryTimeAndCoordinate;
   private final String queryTimeIntervalAndCoordinates;
+  private final String queryTimeKeysAfter;
+  private final String getQueryTimeKeysAfterAndCoordinate;
 
   /**
    * Initializes a new SqlWeatherSource
@@ -88,6 +90,12 @@ public class SqlWeatherSource extends WeatherSource {
             schemaName, weatherTableName, dbTimeColumnName, dbCoordinateIdColumnName);
     this.queryTimeIntervalAndCoordinates =
         createQueryStringForTimeIntervalAndCoordinates(
+            schemaName, weatherTableName, dbTimeColumnName, dbCoordinateIdColumnName);
+
+    this.queryTimeKeysAfter =
+        createQueryStringForTimeKeysAfter(schemaName, weatherTableName, dbTimeColumnName);
+    this.getQueryTimeKeysAfterAndCoordinate =
+        createQueryStringForTimeKeysAfterAndCoordinate(
             schemaName, weatherTableName, dbTimeColumnName, dbCoordinateIdColumnName);
   }
 
@@ -166,6 +174,51 @@ public class SqlWeatherSource extends WeatherSource {
     return Optional.of(timeBasedValues.get(0));
   }
 
+  @Override
+  public Map<Point, List<ZonedDateTime>> getTimeKeysAfter(ZonedDateTime time)
+      throws SourceException {
+
+    Map<Point, Set<TimeBasedValue<WeatherValue>>> timeBasedValues =
+        buildTimeBasedValues(
+                weatherFactory,
+                dataSource.executeQuery(
+                    queryTimeKeysAfter, ps -> ps.setTimestamp(1, Timestamp.from(time.toInstant()))))
+            .stream()
+            .collect(
+                Collectors.groupingBy(
+                    timeBasedWeatherValue -> timeBasedWeatherValue.getValue().getCoordinate(),
+                    Collectors.toSet()));
+
+    return timeBasedValues.entrySet().stream()
+        .collect(
+            Collectors.toMap(
+                Map.Entry::getKey,
+                t -> t.getValue().stream().map(TimeBasedValue::getTime).sorted().toList()));
+  }
+
+  @Override
+  public List<ZonedDateTime> getTimeKeysAfter(ZonedDateTime time, Point coordinate)
+      throws SourceException {
+    Optional<Integer> coordinateId = idCoordinateSource.getId(coordinate);
+    if (coordinateId.isEmpty()) {
+      log.warn("Unable to match coordinate {} to a coordinate ID", coordinate);
+      return Collections.emptyList();
+    }
+
+    return buildTimeBasedValues(
+            weatherFactory,
+            dataSource.executeQuery(
+                getQueryTimeKeysAfterAndCoordinate,
+                ps -> {
+                  ps.setInt(1, coordinateId.get());
+                  ps.setTimestamp(2, Timestamp.from(time.toInstant()));
+                }))
+        .stream()
+        .map(TimeBasedValue::getTime)
+        .sorted()
+        .toList();
+  }
+
   // -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
 
   /**
@@ -184,6 +237,45 @@ public class SqlWeatherSource extends WeatherSource {
         + WHERE
         + timeColumnName
         + " BETWEEN ? AND ?;";
+  }
+
+  /**
+   * Creates a base query to retrieve all time keys after the given time with the following pattern:
+   * <br>
+   * {@code <base query> WHERE <time column> > ?;}
+   *
+   * @param schemaName the name of the database schema
+   * @param weatherTableName the name of the database table
+   * @param timeColumnName the name of the column holding the timestamp info
+   * @return the query string
+   */
+  private static String createQueryStringForTimeKeysAfter(
+      String schemaName, String weatherTableName, String timeColumnName) {
+    return createBaseQueryString(schemaName, weatherTableName) + WHERE + timeColumnName + " > ?;";
+  }
+
+  /**
+   * Creates a basic query to retrieve an entry for the given time and coordinate with the following
+   * pattern: <br>
+   * {@code <base query> WHERE <coordinate column>=? AND <time column> > ?;}
+   *
+   * @param schemaName the name of the database schema
+   * @param weatherTableName the name of the database table
+   * @param timeColumnName the name of the column holding the timestamp info
+   * @param coordinateColumnName name of the column holding the coordinate id
+   * @return the query string
+   */
+  private static String createQueryStringForTimeKeysAfterAndCoordinate(
+      String schemaName,
+      String weatherTableName,
+      String timeColumnName,
+      String coordinateColumnName) {
+    return createBaseQueryString(schemaName, weatherTableName)
+        + WHERE
+        + coordinateColumnName
+        + "=? AND "
+        + timeColumnName
+        + " > ?;";
   }
 
   /**