+++ src/test/groovy/edu/ie3/datamodel/io/source/csv/CsvWeatherSourceCosmoTest.groovy
@@ -17,6 +17,7 @@ import edu.ie3.datamodel.models.value.TemperatureValue
 import edu.ie3.datamodel.models.value.WeatherValue
 import edu.ie3.datamodel.models.value.WindValue
 import edu.ie3.test.common.CosmoWeatherTestData
+import edu.ie3.test.common.IconWeatherTestData
 import edu.ie3.test.helper.WeatherSourceTestHelper
 import edu.ie3.util.TimeUtil
 import edu.ie3.util.geo.GeoUtils
@@ -26,6 +27,8 @@ import org.locationtech.jts.geom.Point
 import spock.lang.Shared
 import spock.lang.Specification
 import tech.units.indriya.quantity.Quantities
+
+import java.time.ZonedDateTime
 
 class CsvWeatherSourceCosmoTest extends Specification implements CsvTestDataMeta, WeatherSourceTestHelper {
 
@@ -224,5 +227,22 @@ class CsvWeatherSourceCosmoTest extends Specification implements CsvTestDataMeta
 
     then:
     actual.empty
+  }
+
+  def "The CsvWeatherSource returns all time keys after a given time key correctly"() {
+    given:
+    def time = TimeUtil.withDefaults.toZonedDateTime("2020-04-28T15:00:00+00:00")
+    def TIME_16H = time.plusHours(1)
+    def TIME_17H = time.plusHours(2)
+
+    when:
+    def actual = source.getTimeKeysAfter(time)
+
+    then:
+    actual.size() == 3
+
+    actual.get(IconWeatherTestData.COORDINATE_193186) == [TIME_16H, TIME_17H]
+    actual.get(IconWeatherTestData.COORDINATE_193187) == [TIME_16H]
+    actual.get(IconWeatherTestData.COORDINATE_193188) == []
   }
 }