+++ src/test/groovy/edu/ie3/datamodel/io/source/csv/CsvWeatherSourceIconTest.groovy
@@ -21,6 +21,8 @@ import org.locationtech.jts.geom.Point
 import spock.lang.Shared
 import spock.lang.Specification
 
+import java.time.ZonedDateTime
+
 class CsvWeatherSourceIconTest extends Specification implements CsvTestDataMeta, WeatherSourceTestHelper {
 
   @Shared
@@ -277,5 +279,22 @@ class CsvWeatherSourceIconTest extends Specification implements CsvTestDataMeta,
 
     then:
     actual.empty
+  }
+
+  def "The CsvWeatherSource returns all time keys after a given time key correctly"() {
+    given:
+    def time = IconWeatherTestData.TIME_15H
+
+    when:
+    def actual = source.getTimeKeysAfter(time)
+
+    then:
+    actual.size() == 2
+
+    actual.get(IconWeatherTestData.COORDINATE_67775) == [
+      IconWeatherTestData.TIME_16H,
+      IconWeatherTestData.TIME_17H
+    ]
+    actual.get(IconWeatherTestData.COORDINATE_67776) == [IconWeatherTestData.TIME_16H]
   }
 }